(Chore) refactor code
