+++ package.json
@@ -10,6 +10,7 @@
     "start": "nodemon --exec babel-node src/estimator.js"
   },
   "dependencies": {
+    "express": "^4.17.1",
     "on-covid-19": "^0.1.3"
   },
   "devDependencies": {
@@ -18,10 +19,12 @@
     "@babel/node": "^7.8.7",
     "@babel/preset-env": "^7.9.0",
     "babel-jest": "^25.2.0",
-    "jest": "^25.2.0",
+    "body-parser": "^1.19.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.20.1",
+    "jest": "^25.2.0",
+    "mongoose": "^5.9.7",
     "nodemon": "^2.0.2"
   },
   "browserslist": [
